requirements.txt
aiohttp
aiosignal
altair
annotated-types
anyio
astroid
async-timeout
attrs
autopep8
beautifulsoup4
blinker
cachetools
certifi
charset-normalizer
click
dataclasses-json
dill
distro
docstring-to-markdown
exceptiongroup
flake8
frozenlist
gitdb
GitPython
greenlet
h11
httpcore
httpx
idna
importlib-metadata
isort
jedi
Jinja2
jsonpatch
jsonpointer
jsonschema
jsonschema-specifications
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langsmith
markdown-it-py
MarkupSafe
marshmallow
mccabe
mdurl
multidict
mypy-extensions
neo4j
numpy
openai
orjson
packaging
pandas
parso
pillow
platformdirs
pluggy
protobuf
pyarrow
pycodestyle
pydantic
pydantic_core
pydeck
pydocstyle
pyflakes
Pygments
pylint
PyMuPDF
PyMuPDFb
python-dateutil
python-dotenv
python-lsp-jsonrpc
python-lsp-server
pytoolconfig
pytz
PyYAML
referencing
regex
requests
requests-toolbelt
rich
rope
rpds-py
six
smmap
sniffio
snowballstemmer
soupsieve
SQLAlchemy
streamlit
tenacity
tiktoken
toml
tomli
tomlkit
toolz
tornado
tqdm
typing-inspect
typing_extensions
tzdata
ujson
urllib3
whatthepatch
yapf
yarl
zipp
